Add files via upload
This is the first script that I wrote in an enterprise setting years ago and I'll never forget it.  The scripts that I use today are lightyears ahead of this one; they're more verbose, more efficient, cover a wide array of tasks; but they all started with this one batch script.  The aim is to build on this script to show the nature progression of curiosity and how it positively impacted my job role.  I hope you'll be inspired!